Fix Duplicate Provider Bug #1391
Fluent button has both "name" and "Value".  When the form is submitted these two values are concatenated together with a comma. 

Ex: "Google" becomes "Google,Google" and the login fails because there is no matching provider.

This was copied directly from the default Blazor template and the FluentButton seems to behave differently than a stock html button.