[Stack] Added 'Stretch' horizontal alignment option (#2800)
* Added missing 'Stretch' horizontal alignment option to FluentStack

Introduced a new horizontal alignment option, `Stretch`, to the `FluentStack` component. Updated the `HorizontalAlignment` enum in `HorizontalAlignment.cs` to include the `Stretch` value with a summary comment. Modified the `GetHorizontalAlignment` method in `FluentStack.razor.cs` to map `HorizontalAlignment.Stretch`  to the string "stretch".

* Add validation to Alignment property in DialogParameters

The Alignment property now includes a custom setter that throws an ArgumentException if set to HorizontalAlignment.Stretch, ensuring this value is not supported.

* Moved "_alignment" variable declaration to top of DialogParameters class

* Added a note re. stretch to the XML documentation for the `Alignment` property

* Added unit test for DialogParameters invalid Aligment value

---------

Co-authored-by: Vincent Baaij <vnbaaij@outlook.com>
Co-authored-by: Denis Voituron <dvoituron@outlook.com>